compiler_gym/requirements.txt
absl-py>=0.10.0
deprecated>=1.2.12
fasteners>=0.15
grpcio>=1.32.0
gym>=0.18.0
humanize>=2.6.0
networkx>=2.5
numpy>=1.19.3
protobuf>=3.13.0
pydantic>=1.8.0
requests>=2.24.0
tabulate>=0.8.7
